add block placement

+++ include/GameData/Block.hpp
@@ -9,12 +9,12 @@ struct BlockState;
 struct Block {
     BlockId type;
     BlockState* state;
-    glm::vec3 position;
+    glm::ivec3 position;
 
     bool isSolid() const;
     bool isTransparent() const;
 
     Block();
-    Block(const glm::vec3& position, BlockId type);
-    Block(const glm::vec3& position, BlockId type, BlockState* state);
+    Block(const glm::ivec3& position, BlockId type);
+    Block(const glm::ivec3& position, BlockId type, BlockState* state);
 };
